@@ -4,20 +4,21 @@ version: 0.1.7
 homepage: https://github.com/andrey-ushakov/flutter_bluetooth_basic
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.18.6 <3.0.0'
   flutter: ">=1.12.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  json_annotation: ^4.1.0
-  rxdart: ^0.26.0
+  json_annotation: ^4.7.0
+
+  rxdart: ^0.27.7
 
 dev_dependencies:
+  build_runner: ^2.3.3
   flutter_test:
     sdk: flutter
-  build_runner: ^1.0.0
-  json_serializable: ^3.2.2
+  json_serializable: ^6.5.4
 
 flutter:
   plugin:
